+++ packages/react-ui/components/Kebab/Kebab.styles.ts
@@ -73,7 +73,7 @@ const styles = {
   iconlarge() {
     return css`
       ${cssName(styles.kebab())} & {
-        margin-top: -5px;
+        margin-top: -4px;
         margin-left: 0.5px;
       }
     `;