+++ packages/react-ui/components/Spinner/Spinner.styles.ts
@@ -1,23 +1,32 @@
 import { css, memoizeStyle } from '../../lib/theming/Emotion';
 import { Theme } from '../../lib/theming/Theme';
 import { AnimationKeyframes } from '../../lib/theming/AnimationKeyframes';
+import { isIE11 } from '../../lib/utils';
 
 const styles = {
   circle(t: Theme) {
     return css`
       stroke: ${t.spinnerColor};
-      animation: ${AnimationKeyframes.spinnerCircleOffset(t)} 1s cubic-bezier(0.5, 0.2, 0.5, 0.8) infinite,
-        ${AnimationKeyframes.spinnerCircleLength(t)} 2s cubic-bezier(0.36, 0.14, 0.38, 0.69) infinite,
-        ${AnimationKeyframes.spinnerCircleRotate(t)} 2s linear infinite,
-        ${AnimationKeyframes.spinnerColor(t)} 6s ease-in-out infinite;
+
+      ${!isIE11 &&
+        `
+          animation: ${AnimationKeyframes.spinnerCircleOffset(t)} 1s cubic-bezier(0.5, 0.2, 0.5, 0.8) infinite,
+            ${AnimationKeyframes.spinnerCircleLength(t)} 2s cubic-bezier(0.36, 0.14, 0.38, 0.69) infinite,
+            ${AnimationKeyframes.spinnerCircleRotate(t)} 2s linear infinite,
+            ${AnimationKeyframes.spinnerColor(t)} 6s ease-in-out infinite;
+        `}
     `;
   },
   circleDimmed(t: Theme) {
     return css`
       stroke: ${t.spinnerDimmedColor};
-      animation: ${AnimationKeyframes.spinnerCircleOffset(t)} 1s cubic-bezier(0.5, 0.2, 0.5, 0.8) infinite,
-        ${AnimationKeyframes.spinnerCircleLength(t)} 2s cubic-bezier(0.36, 0.14, 0.38, 0.69) infinite,
-        ${AnimationKeyframes.spinnerCircleRotate(t)} 2s linear infinite;
+
+      ${!isIE11 &&
+        `
+          animation: ${AnimationKeyframes.spinnerCircleOffset(t)} 1s cubic-bezier(0.5, 0.2, 0.5, 0.8) infinite,
+          ${AnimationKeyframes.spinnerCircleLength(t)} 2s cubic-bezier(0.36, 0.14, 0.38, 0.69) infinite,
+          ${AnimationKeyframes.spinnerCircleRotate(t)} 2s linear infinite;
+        `}
     `;
   },
 
@@ -63,13 +72,6 @@ const styles = {
   inner() {
     return css`
       display: inline-block;
-    `;
-  },
-
-  fallback() {
-    return css`
-      display: inline-block;
-      position: relative;
     `;
   },
 };