+++ packages/react-ui/internal/SpinnerOld/SpinnerOld.styles.ts
@@ -1,6 +1,7 @@
 import { css, memoizeStyle } from '../../lib/theming/Emotion';
 import { Theme } from '../../lib/theming/Theme';
 import { AnimationKeyframes } from '../../lib/theming/AnimationKeyframes';
+import { isIE11 } from '../../lib/utils';
 
 const styles = {
   cloudStroke(t: Theme) {
@@ -19,9 +20,13 @@ const styles = {
       stroke-dasharray: 20, 86;
       stroke-dashoffset: 15;
       stroke: ${t.spinnerColor};
-      animation: ${AnimationKeyframes.spinnerCloudOffset(t)} 1s cubic-bezier(0.5, 0.2, 0.5, 0.8) infinite,
-        ${AnimationKeyframes.spinnerCloudLength(t)} 2s cubic-bezier(0.36, 0.14, 0.38, 0.69) infinite,
-        ${AnimationKeyframes.spinnerColor(t)} 6s ease-in-out infinite;
+
+      ${!isIE11 &&
+        `
+        animation: ${AnimationKeyframes.spinnerCloudOffset(t)} 1s cubic-bezier(0.5, 0.2, 0.5, 0.8) infinite,
+          ${AnimationKeyframes.spinnerCloudLength(t)} 2s cubic-bezier(0.36, 0.14, 0.38, 0.69) infinite,
+          ${AnimationKeyframes.spinnerColor(t)} 6s ease-in-out infinite;
+        `}
     `;
   },
   cloudDimmed(t: Theme) {
@@ -29,25 +34,37 @@ const styles = {
       stroke-dasharray: 20, 86;
       stroke-dashoffset: 15;
       stroke: ${t.spinnerOldDimmedColor};
-      animation: ${AnimationKeyframes.spinnerCloudOffset(t)} 1s cubic-bezier(0.5, 0.2, 0.5, 0.8) infinite,
-        ${AnimationKeyframes.spinnerCloudLength(t)} 2s cubic-bezier(0.36, 0.14, 0.38, 0.69) infinite;
+
+      ${!isIE11 &&
+        `
+        animation: ${AnimationKeyframes.spinnerCloudOffset(t)} 1s cubic-bezier(0.5, 0.2, 0.5, 0.8) infinite,
+          ${AnimationKeyframes.spinnerCloudLength(t)} 2s cubic-bezier(0.36, 0.14, 0.38, 0.69) infinite;
+        `}
     `;
   },
   circle(t: Theme) {
     return css`
       stroke: ${t.spinnerOldColor};
-      animation: ${AnimationKeyframes.spinnerCircleOffset(t)} 1s cubic-bezier(0.5, 0.2, 0.5, 0.8) infinite,
-        ${AnimationKeyframes.spinnerCircleLength(t)} 2s cubic-bezier(0.36, 0.14, 0.38, 0.69) infinite,
-        ${AnimationKeyframes.spinnerCircleRotate(t)} 2s linear infinite,
-        ${AnimationKeyframes.spinnerColor(t)} 6s ease-in-out infinite;
+
+      ${!isIE11 &&
+        `
+        animation: ${AnimationKeyframes.spinnerCircleOffset(t)} 1s cubic-bezier(0.5, 0.2, 0.5, 0.8) infinite,
+          ${AnimationKeyframes.spinnerCircleLength(t)} 2s cubic-bezier(0.36, 0.14, 0.38, 0.69) infinite,
+          ${AnimationKeyframes.spinnerCircleRotate(t)} 2s linear infinite,
+          ${AnimationKeyframes.spinnerColor(t)} 6s ease-in-out infinite;
+        `}
     `;
   },
   circleDimmed(t: Theme) {
     return css`
       stroke: ${t.spinnerOldDimmedColor};
-      animation: ${AnimationKeyframes.spinnerCircleOffset(t)} 1s cubic-bezier(0.5, 0.2, 0.5, 0.8) infinite,
-        ${AnimationKeyframes.spinnerCircleLength(t)} 2s cubic-bezier(0.36, 0.14, 0.38, 0.69) infinite,
-        ${AnimationKeyframes.spinnerCircleRotate(t)} 2s linear infinite;
+
+      ${!isIE11 &&
+        `
+        animation: ${AnimationKeyframes.spinnerCircleOffset(t)} 1s cubic-bezier(0.5, 0.2, 0.5, 0.8) infinite,
+          ${AnimationKeyframes.spinnerCircleLength(t)} 2s cubic-bezier(0.36, 0.14, 0.38, 0.69) infinite,
+          ${AnimationKeyframes.spinnerCircleRotate(t)} 2s linear infinite;
+        `}
     `;
   },
 
@@ -93,13 +110,6 @@ const styles = {
       font-size: ${t.spinnerOldFontSizeLarge};
       line-height: ${t.spinnerOldLineHeightLarge};
       margin-top: ${t.spinnerOldCaptionGapLarge};
-    `;
-  },
-
-  fallback() {
-    return css`
-      display: inline-block;
-      position: relative;
     `;
   },
 };