+++ packages/react-ui/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.10.0](https://github.com/skbkontur/retail-ui/compare/@skbkontur/react-ui@2.9.0...@skbkontur/react-ui@2.10.0) (2020-12-03)
+
+
+### Bug Fixes
+
+* **Icon20px,TopBar:** replace iconic font with svg icons ([67981e1](https://github.com/skbkontur/retail-ui/commit/67981e125a4e5fc93eb4f5e03c0ebcdfb638b6ed)), closes [#2223](https://github.com/skbkontur/retail-ui/issues/2223)
+* **Input:** fix disabled color in Safari ([5b35688](https://github.com/skbkontur/retail-ui/commit/5b356880aea96fbe716eaf3cb8d0cedf72f51542))
+* **Kebab:** use svg icon for large size ([c3d5b13](https://github.com/skbkontur/retail-ui/commit/c3d5b138ddcf1da45f487616fa940f7ac2d7e921)), closes [#2223](https://github.com/skbkontur/retail-ui/issues/2223)
+* **Modal:** fix cross's box-sizing ([e734022](https://github.com/skbkontur/retail-ui/commit/e7340220f6bfabaed5c5b74fd094767a47c8c4e0))
+* **Spinner,SpinnerOld:** remake fallback animation using SVG  ([#2203](https://github.com/skbkontur/retail-ui/issues/2203)) ([fe32192](https://github.com/skbkontur/retail-ui/commit/fe321928785a464dd2c631266ab18ae508436caf)), closes [#2182](https://github.com/skbkontur/retail-ui/issues/2182)
+* **Textarea:** correct onCut and onPaste handlers ([ee04b65](https://github.com/skbkontur/retail-ui/commit/ee04b65f6b0421df06f08014de6a226c428cb49c))
+
+
+### Features
+
+* **ComboBox:** add leftIcon prop ([d7a4904](https://github.com/skbkontur/retail-ui/commit/d7a49047717a2a5acefdc2136e3edccc0a56a703))
+* **TopBarItem:** accept ReactElement as icon ([4fa4609](https://github.com/skbkontur/retail-ui/commit/4fa460994a0e1043cddfbb1f485cb761b801f3e2))
+
+
+
+
+
 # [2.9.0](https://github.com/skbkontur/retail-ui/compare/@skbkontur/react-ui@2.8.0...@skbkontur/react-ui@2.9.0) (2020-11-19)